Move symlinksSupported to test/util.go
It's used by both symlink_test.go and ignore_test.go and we should be
able to 'go test' ignore_test.go without symlink_test.go.
